When ctrl-c is received try to stop all containers
This tries to stop all containers ever started by this process via this
crate. On the one hand it's an overkill, on the other such approach
simplifies code - no need to keep track of when container dies.

Also this is probably not correct when taken into account requirements
for signal handlers (at least on linux). Maybe this should be enable
under a feature flag?